Touch screen support: Double Tap and Long Tap Support for TraceLayout controls. With this long tapping the clock will show the next overlay, double tapping the clock will show the tacho

+++ GpsMidGraph/de/ueller/gps/tools/IconActionPerformer.java
@@ -1,7 +1,7 @@
 package de.ueller.gps.tools;
 
 public interface IconActionPerformer {
-	public final static int BACK_ACTIONID = Integer.MAX_VALUE;
+	public final static int BACK_ACTIONID = Byte.MAX_VALUE;
 	
 	public void performIconAction(int actionId);
